Partner Portal: Add an empty view for the license list (#51762)
* Partner Portal: Add isolated component for licenses list header

* Partner Portal: Add license list empty view

* Partner Portal: Using existent translation for the empty view

* Partner Portal: Fix license list empty view titles
